# docs/swagger.yaml
basePath: /signpdf
definitions:
  dto.PostSignPdfConfigB:
    properties:
      customSignatureText:
        maxLength: 80
        type: string
      dest:
        maxLength: 300
        minLength: 1
        type: string
      dtsFlag:
        default: 1
        maximum: 1
        minimum: 1
        type: integer
      fileServerId:
        maxLength: 20
        minLength: 1
        type: string
      orgId:
        maxLength: 45
        minLength: 1
        type: string
      pCode:
        maxLength: 45
        minLength: 1
        type: string
      pin:
        maxLength: 16
        minLength: 1
        type: string
      qrPosition:
        maxLength: 300
        type: string
      qrSize:
        default: 0
        maximum: 3
        type: integer
      qrValue:
        maxLength: 200
        type: string
      remark:
        maxLength: 300
        type: string
      signerFontSize:
        maximum: 3
        type: integer
      signerImagePath:
        maxLength: 300
        type: string
      signerTextRow:
        maximum: 3
        type: integer
      signerXyPage:
        maxLength: 300
        minLength: 1
        type: string
      source:
        maxLength: 300
        minLength: 1
        type: string
      userId:
        maxLength: 20
        minLength: 1
        type: string
    required:
    - dest
    - dtsFlag
    - fileServerId
    - orgId
    - pCode
    - pin
    - signerXyPage
    - source
    - userId
    type: object
  dto.PostVerifyRoamingCert:
    properties:
      orgId:
        maxLength: 45
        minLength: 1
        type: string
      pCode:
        maxLength: 45
        minLength: 1
        type: string
      userId:
        maxLength: 20
        minLength: 1
        type: string
    required:
    - orgId
    - pCode
    - userId
    type: object
  dto.PostVerifyRoamingPin:
    properties:
      orgId:
        maxLength: 45
        minLength: 1
        type: string
      pCode:
        maxLength: 45
        minLength: 1
        type: string
      pin:
        maxLength: 16
        minLength: 1
        type: string
      userId:
        maxLength: 20
        minLength: 1
        type: string
    required:
    - orgId
    - pCode
    - pin
    - userId
    type: object
info:
  contact: {}
  description: This is a sign pdf server.
  title: Swagger Sign PDF API
  version: "1.0"
paths:
...
  /dss/signPdf:
    post:
      consumes:
      - application/json
      parameters:
      - description: SignPdfConfigB DTO
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PostSignPdfConfigB'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - DSS
  /dss/verifyRoamingCert:
    post:
      consumes:
      - application/json
      parameters:
      - description: VerifyRoamingCert DTO
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PostVerifyRoamingCert'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - DSS
  /dss/verifyRoamingPin:
    post:
      consumes:
      - application/json
      parameters:
      - description: VerifyRoamingPin DTO
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PostVerifyRoamingPin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - DSS
...
swagger: "2.0"